+++ lib/camera_macos_arguments.dart
@@ -1,7 +1,10 @@
+import 'dart:typed_data';
+
 import 'package:camera_macos/camera_macos_device.dart';
 import 'package:flutter/material.dart';
 
 enum PictureFormat{jpg,jpeg,tiff,bmp,png}
+enum VideoFormat{m4v,mov,mp4}
 enum PictureResolution{
   /// 480p (640x480)
   low,
@@ -20,6 +23,26 @@ enum PictureResolution{
 
   /// The highest resolution available.
   max,
+}
+
+class CameraImageData{
+  CameraImageData({
+    required this.width,
+    required this.height,
+    required this.bytes
+  });
+
+  final int width;
+  final int height;
+  final Uint8List bytes;
+
+  @override
+  String toString(){
+    return {
+      'width': width,
+      'height': height
+    }.toString();
+  }
 }
 
 class CameraMacOSArguments {