+++ macos/Classes/CameraMacosPlugin.swift
@@ -2,10 +2,12 @@ import Cocoa
 import FlutterMacOS
 import AVFoundation
 
-public class CameraMacosPlugin: NSObject, FlutterPlugin, FlutterTexture, AVCaptureVideoDataOutputSampleBufferDelegate, AVCaptureAudioDataOutputSampleBufferDelegate, AVAssetWriterDelegate, AVCaptureFileOutputRecordingDelegate {
+public class CameraMacosPlugin: NSObject, FlutterPlugin, FlutterTexture, AVCaptureVideoDataOutputSampleBufferDelegate, AVCaptureAudioDataOutputSampleBufferDelegate, AVAssetWriterDelegate, AVCaptureFileOutputRecordingDelegate,
+    FlutterStreamHandler{
     
     let registry: FlutterTextureRegistry
     let outputChannel: FlutterMethodChannel!
+    var sink: FlutterEventSink!
     
     // Texture id of the camera preview
     var textureId: Int64!
@@ -49,7 +51,9 @@ public class CameraMacosPlugin: NSObject, FlutterPlugin, FlutterTexture, AVCaptu
     var factory: CameraMacOSNativeFactory!
     var previewLayer: AVCaptureVideoPreviewLayer!
 
-    var format:NSBitmapImageRep.FileType = NSBitmapImageRep.FileType.tiff
+    var pictureFormat:NSBitmapImageRep.FileType = NSBitmapImageRep.FileType.tiff
+    var videoFormat:AVFileType = AVFileType.mp4
+    var vstring:String = "mp4"
     var resSize:NSSize? = nil
     var settingsAssistant:AVOutputSettingsAssistant? = nil
     
@@ -57,6 +61,17 @@ public class CameraMacosPlugin: NSObject, FlutterPlugin, FlutterTexture, AVCaptu
         self.registry = registry
         self.outputChannel = outputChannel
         super.init()
+    }
+    // FlutterStreamHandler
+    public func onListen(withArguments arguments: Any?, eventSink events: @escaping FlutterEventSink) -> FlutterError? {
+        sink = events
+        return nil
+    }
+    
+    // FlutterStreamHandler
+    public func onCancel(withArguments arguments: Any?) -> FlutterError? {
+        sink = nil
+        return nil
     }
     
     public static func register(with registrar: FlutterPluginRegistrar) {
@@ -67,11 +82,30 @@ public class CameraMacosPlugin: NSObject, FlutterPlugin, FlutterTexture, AVCaptu
         let factory = CameraMacOSNativeFactory(messenger: registrar.messenger)
         registrar.register(factory, withId: "camera_macos_view")
         instance.factory = factory
+        
+        // Channel for communicating with platform plugins using event streams
+        let event = FlutterEventChannel(name:"camera_macos/stream", binaryMessenger: registrar.messenger)
+        registrar.addMethodCallDelegate(instance, channel: outputChannel)
+        event.setStreamHandler(instance)
     }
     
     public func copyPixelBuffer() -> Unmanaged<CVPixelBuffer>? {
         if latestBuffer == nil {
             return nil
+        }
+        if self.sink != nil{
+            let u = imageFromSampleBuffer(imageBuffer: latestBuffer)!
+            let bytesPerRow = u.bytesPerRow
+            let width = Int(u.size.width)
+            let height = Int(u.size.height)
+            
+            let newData:Data = Data(bytes: u.bitmapData!, count: Int(bytesPerRow*height))
+            
+            self.sink([
+                "width": width,
+                "height": height,
+                "data": newData,
+            ]);
         }
         return Unmanaged<CVPixelBuffer>.passRetained(latestBuffer)
     }
@@ -89,7 +123,7 @@ public class CameraMacosPlugin: NSObject, FlutterPlugin, FlutterTexture, AVCaptu
             }
             initCamera(arguments, result)
         case "takePicture":
-            takePicture(result,format)
+            takePicture(result,pictureFormat)
         case "startRecording":
             let arguments = call.arguments as? Dictionary<String, Any> ?? [:]
             startRecording(arguments, result)
@@ -97,11 +131,39 @@ public class CameraMacosPlugin: NSObject, FlutterPlugin, FlutterTexture, AVCaptu
             stopRecording(result)
         case "destroy":
             destroy(result)
+        case "setFocusPoint":
+            guard let arguments = call.arguments as? Dictionary<String, Any>
+            else {
+                result(FlutterError(code: "INVALID_ARGS", message: "", details: nil).toMap)
+                return
+            }
+            setFocusPoint(arguments, result)
         default:
             result(FlutterMethodNotImplemented)
         }
     }
-    
+
+    func setFocusPoint(_ arguments: Dictionary<String, Any>, _ result: @escaping FlutterResult) {
+        var capturedVideoDevices: [AVCaptureDevice] = []
+        
+        if #available(macOS 10.15, *) {
+            capturedVideoDevices = AVCaptureDevice.captureDevices(deviceTypes: [.builtInWideAngleCamera, .externalUnknown], mediaType: .video)
+        } else {
+            capturedVideoDevices = AVCaptureDevice.captureDevices(mediaType: .video)
+        }
+
+        let deviceId = arguments["deviceType"] as? String;
+        if let newCameraObject = capturedVideoDevices.first(where: { $0.uniqueID == deviceId }) {
+            let x = arguments["x"] as! Double;
+            let y = arguments["y"] as! Double;
+
+            let focusPoint: CGPoint = .init(x: x, y: y)
+            if newCameraObject.isFocusPointOfInterestSupported {
+                newCameraObject.focusPointOfInterest = focusPoint
+            }
+        }
+    }
+
     func requestPermission(completionHandler: @escaping (Bool) -> Void) {
         if #available(macOS 10.14, *) {
             AVCaptureDevice.requestAccess(for: .video, completionHandler: completionHandler)
@@ -112,9 +174,9 @@ public class CameraMacosPlugin: NSObject, FlutterPlugin, FlutterTexture, AVCaptu
     
     func generateVideoFileURL(randomGUID: Bool = false) -> URL {
         let paths = FileManager.default.urls(for: .cachesDirectory, in: .userDomainMask)
-        var fileUrl = paths[0].appendingPathComponent("output.mp4")
+        var fileUrl = paths[0].appendingPathComponent("output."+vstring)
         if randomGUID {
-            fileUrl = paths[0].appendingPathComponent(UUID().uuidString + ".mp4")
+            fileUrl = paths[0].appendingPathComponent(UUID().uuidString + "."+vstring)
         }
         return fileUrl
     }
@@ -151,7 +213,7 @@ public class CameraMacosPlugin: NSObject, FlutterPlugin, FlutterTexture, AVCaptu
             }
         }
     }
-    
+
     func initCamera(_ arguments: Dictionary<String, Any>, _ result: @escaping FlutterResult) {
         if let captureSession = self.captureSession, captureSession.isRunning {
             captureSession.stopRunning()
@@ -203,21 +265,35 @@ public class CameraMacosPlugin: NSObject, FlutterPlugin, FlutterTexture, AVCaptu
                     newCameraObject = capturedVideoDevices.first
                 }
                 
-                switch arguments["format"] as! String{
+                switch arguments["pformat"] as! String{
                     case "jpg":
-                        self.format = NSBitmapImageRep.FileType.jpeg
+                        self.pictureFormat = NSBitmapImageRep.FileType.jpeg
                         break
                     case "jepg":
-                        self.format = NSBitmapImageRep.FileType.jpeg2000
+                        self.pictureFormat = NSBitmapImageRep.FileType.jpeg2000
                         break
                     case "bmp":
-                        self.format = NSBitmapImageRep.FileType.bmp
+                        self.pictureFormat = NSBitmapImageRep.FileType.bmp
                         break
                     case "png":
-                        self.format = NSBitmapImageRep.FileType.png
+                        self.pictureFormat = NSBitmapImageRep.FileType.png
                         break
                     default:
-                        self.format = NSBitmapImageRep.FileType.tiff
+                        self.pictureFormat = NSBitmapImageRep.FileType.tiff
+                        break
+                }
+                switch arguments["vformat"] as! String{
+                    case "m4v":
+                        self.videoFormat = AVFileType.m4v
+                        self.vstring = "m4v"
+                        break
+                    case "mov":
+                        self.videoFormat = AVFileType.mov
+                        self.vstring = "mov"
+                        break
+                    default:
+                        self.videoFormat = AVFileType.mp4
+                        self.vstring = "mp4"
                         break
                 }
                 switch arguments["resolution"] as! String{
@@ -305,7 +381,7 @@ public class CameraMacosPlugin: NSObject, FlutterPlugin, FlutterTexture, AVCaptu
                     var outputInitialized: Bool = false
                     
                     self.useMovieFileOutput = arguments["useMovieFileOutput"] as? Bool ?? false
-//                    #warning("forced to use MovieFileOutput for testing purposes")
+                    // #warning("forced to use MovieFileOutput for testing purposes")
                     
                     // Add video buffering output
                     
@@ -451,15 +527,13 @@ public class CameraMacosPlugin: NSObject, FlutterPlugin, FlutterTexture, AVCaptu
             return image
         }
             
-        var newImage = NSImage(size: resSize!)
+        let newImage = NSImage(size: resSize!)
         newImage.lockFocus()
         image.draw(in: NSMakeRect(0, 0, resSize!.width, resSize!.height), from: NSMakeRect(0, 0, image.size.width, image.size.height), operation: NSCompositingOperation.sourceOver, fraction: CGFloat(1), respectFlipped: false, hints: nil)
             
         newImage.unlockFocus()
         newImage.size = resSize!
         return NSBitmapImageRep(data: newImage.tiffRepresentation!)
-        
-        //return (image);
     }
     
     func startRecording(_ arguments: Dictionary<String, Any>, _ result: @escaping FlutterResult) {
@@ -500,7 +574,7 @@ public class CameraMacosPlugin: NSObject, FlutterPlugin, FlutterTexture, AVCaptu
                     self.savedResult = result
                     movieOutput.startRecording(to: fileUrl, recordingDelegate: self)
                 } else {
-                    self.videoWriter = try AVAssetWriter(outputURL: fileUrl, fileType: .mp4)
+                    self.videoWriter = try AVAssetWriter(outputURL: fileUrl, fileType: videoFormat)
                     print("Setting up AVAssetWriter")
 
                     guard let videoWriter = self.videoWriter else {